fix(cli): store phase physics as f64 in Config so the template serializes cleanly
The phase physics Config fields were f32, so `tomoxide init` wrote decimal
quantities with f32→f64 promotion noise (`pixel_size = 0.00009999999747378752`
instead of `0.0001`) — confusing in a file users read and edit. Store them as f64
(the natural type for physical measurements) and cast to the f32 the phase methods
use at the resolve() boundary. Template now reads `0.0001`/`0.001`/`0.0002`;
custom config values still load and cast correctly (verified pixel_size=0.00075,
energy=22.5 round-trip). clippy -p tomoxide-cli (default + cuda) clean.

+++ crates/tomoxide-cli/src/config.rs
@@ -60,18 +60,21 @@ pub struct Config {
     pub vo_sm_size: usize,
 
     // --- Phase-retrieval physics (used when a phase method is selected) ---
+    // Stored as f64 so decimal quantities like `1e-4` serialize cleanly in the
+    // template (an f32 field promotes to f64 on write and leaks precision noise,
+    // e.g. `0.00009999999747…`); cast to f32 at the reconstruction boundary.
     /// Detector pixel size (cm).
-    pub pixel_size: f32,
+    pub pixel_size: f64,
     /// Sample-to-detector propagation distance (cm).
-    pub propagation_distance: f32,
+    pub propagation_distance: f64,
     /// X-ray energy (keV).
-    pub energy: f32,
+    pub energy: f64,
     /// Paganin regularization parameter `alpha`.
-    pub alpha: f32,
+    pub alpha: f64,
     /// Gpaganin/farago material `delta/beta` ratio.
-    pub db: f32,
+    pub db: f64,
     /// Gpaganin characteristic transverse length `W` (cm).
-    pub w: f32,
+    pub w: f64,
 }
 
 impl Default for Config {
